requirements.txt
click==6.6
coreapi==2.0.7
coreapi-cli==1.0.5
Django==1.10.1
django-cors-middleware==1.3.1
djangorestframework==3.4.7
gunicorn==19.6.0
itypes==1.1.0
mysqlclient==1.3.7
psycopg2==2.6.2
requests==2.11.1
uritemplate==3.0.0
